change button to "Post Say"

+++ CharmChart/CharmChart.m
@@ -323,7 +323,7 @@
 }
 - (IBAction)btnCloseClicked:(id)sender withCharm:(NSString*)selectedCharm{
     if([self.delegate respondsToSelector:@selector(showCharmsSelection: withIndex:)]) {
-        [self.delegate showCharmsSelection:_title withIndex:[NSString stringWithFormat:@"%i", _index]];
+        [self.delegate showCharmsSelection:_title withIndex:[NSString stringWithFormat:@"%li", (long)_index]];
     }
 
 }
